Prototype 1: Proposal for addressing weakness in entropy pooling.
There is an issue in how entropy is being preserved in a source's pool.
The initial implementation attempted to perform incremental hashing as
a means to generate a progressive hash of incoming entropy.  This does
not work as intended, as the .NET API does not expose the capability to
calculate hashes of incremental blocks of data.

This solution attempts to obviate the need for storing large strings of
entropic data for rehashing, and instead attempts to seed each hashing
invocation with the results of the previous hashing computation.